+++ src/relstorage/adapters/mysql/mover.py
@@ -162,13 +162,11 @@ class MySQLObjectMover(AbstractObjectMover):
     """
 
     # UPSERT for current_object: no need for separate update.
-    _update_current_insert_query = """
-        INSERT INTO current_object (zoid, tid)
-            SELECT zoid, tid FROM object_state
-            WHERE tid = %s
-            ORDER BY zoid
+    _update_current_insert_query = (
+        AbstractObjectMover._upsert_current_insert_base
+        + """
         ON DUPLICATE KEY UPDATE
             tid = VALUES(tid)
-    """
-
+        """
+    )
     _update_current_update_query = None